+++ source/fuzz/fuzzer.cpp
@@ -82,13 +82,13 @@
 #include "source/fuzz/fuzzer_pass_swap_commutable_operands.h"
 #include "source/fuzz/fuzzer_pass_swap_conditional_branch_operands.h"
 #include "source/fuzz/fuzzer_pass_toggle_access_chain_instruction.h"
+#include "source/fuzz/pass_management/repeated_pass_manager.h"
+#include "source/fuzz/pass_management/repeated_pass_manager_looped_with_recommendations.h"
+#include "source/fuzz/pass_management/repeated_pass_manager_random_with_recommendations.h"
+#include "source/fuzz/pass_management/repeated_pass_manager_simple.h"
+#include "source/fuzz/pass_management/repeated_pass_recommender_standard.h"
 #include "source/fuzz/protobufs/spirvfuzz_protobufs.h"
 #include "source/fuzz/pseudo_random_generator.h"
-#include "source/fuzz/repeated_pass_manager.h"
-#include "source/fuzz/repeated_pass_manager_looped_with_recommendations.h"
-#include "source/fuzz/repeated_pass_manager_random_with_recommendations.h"
-#include "source/fuzz/repeated_pass_manager_simple.h"
-#include "source/fuzz/repeated_pass_recommender_standard.h"
 #include "source/fuzz/transformation_context.h"
 #include "source/opt/build_module.h"
 #include "source/spirv_fuzzer_options.h"
