+++ source/fuzz/pass_management/repeated_pass_manager.cpp
@@ -12,27 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SOURCE_FUZZ_REPEATED_PASS_MANAGER_SIMPLE_H_
-#define SOURCE_FUZZ_REPEATED_PASS_MANAGER_SIMPLE_H_
-
-#include "source/fuzz/repeated_pass_manager.h"
+#include "source/fuzz/pass_management/repeated_pass_manager.h"
 
 namespace spvtools {
 namespace fuzz {
 
-// Selects the next pass to run uniformly at random from the enabled repeated
-// passes.  Recommendations are not used.
-class RepeatedPassManagerSimple : public RepeatedPassManager {
- public:
-  RepeatedPassManagerSimple(FuzzerContext* fuzzer_context,
-                            RepeatedPassInstances* pass_instances);
+RepeatedPassManager::RepeatedPassManager(FuzzerContext* fuzzer_context,
+                                         RepeatedPassInstances* pass_instances)
+    : fuzzer_context_(fuzzer_context), pass_instances_(pass_instances) {}
 
-  ~RepeatedPassManagerSimple() override;
-
-  FuzzerPass* ChoosePass() override;
-};
+RepeatedPassManager::~RepeatedPassManager() = default;
 
 }  // namespace fuzz
 }  // namespace spvtools
-
-#endif  // SOURCE_FUZZ_REPEATED_PASS_MANAGER_SIMPLE_H_
