+++ source/fuzz/pass_management/repeated_pass_manager.h
@@ -17,7 +17,7 @@
 
 #include "source/fuzz/fuzzer_context.h"
 #include "source/fuzz/fuzzer_pass.h"
-#include "source/fuzz/repeated_pass_instances.h"
+#include "source/fuzz/pass_management/repeated_pass_instances.h"
 
 namespace spvtools {
 namespace fuzz {
