+++ source/fuzz/pass_management/repeated_pass_manager_simple.cpp
@@ -12,12 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "source/fuzz/repeated_pass_recommender.h"
+#include "source/fuzz/pass_management/repeated_pass_manager_simple.h"
 
 namespace spvtools {
 namespace fuzz {
 
-RepeatedPassRecommender::~RepeatedPassRecommender() = default;
+RepeatedPassManagerSimple::RepeatedPassManagerSimple(
+    FuzzerContext* fuzzer_context, RepeatedPassInstances* pass_instances)
+    : RepeatedPassManager(fuzzer_context, pass_instances) {}
+
+RepeatedPassManagerSimple::~RepeatedPassManagerSimple() = default;
+
+FuzzerPass* RepeatedPassManagerSimple::ChoosePass() {
+  auto& passes = GetPassInstances()->GetPasses();
+  return passes[GetFuzzerContext()->RandomIndex(passes)].get();
+}
 
 }  // namespace fuzz
 }  // namespace spvtools
